move test fixture
